feat: code refactoring
Dropped unused code & restructured everything.
