+++ src/components/Block/ResourcesTeaser.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import Link from "next/link";
 import { IoLogoGithub, IoLogoYoutube, IoMdChatbubbles } from "react-icons/io";
 
-export default function BlockResources() {
+export default function BlockResourcesTeaser() {
   return (
     <div className="uiBlock darkBackground">
       <Container>
