+++ src/content/news.json
@@ -1,0 +1,86 @@
+[
+  {
+    "name": "ERC-4626 Proposes To Make Yield-Bearing Tokens Their Own DeFi Lego",
+    "href": "https://thedefiant.io/erc-4626-yield-standard/",
+    "date": "2022-01-06",
+    "thumbnail": "/images/thedefiant-preview.png",
+    "teaser": "A new proposed token standard for Ethereum sent optimistic ripples through the DeFi community this week. "
+  },
+  {
+    "name": "ERC-4626: DeFi's Newest Money Lego",
+    "teaser": "An Ethereum Improvement Proposal wants to standardize a key component in yield-generating strategies.",
+    "href": "https://www.coindesk.com/layer2/2022/01/13/erc-4626-defis-newest-money-lego/",
+    "date": "2022-01-13",
+    "thumbnail": "https://www.coindesk.com/resizer/5BzHsoS9e0btkKAjjwZsBG3Y9I0=/1200x628/center/middle/cloudfront-us-east-1.images.arcpublishing.com/coindesk/445CGEBDLJEHTI6OFLZJN3YCHA.png"
+  },
+  {
+    "name": "DeFi Giant Yearn Leads the Way on ERC-4626 Token Standard Adoption",
+    "teaser": "The new standard for yield-bearing tokens could unlock innovation in the burgeoning Ethereum economy.",
+    "href": "https://www.coindesk.com/layer2/2022/04/08/defi-giant-yearn-leads-the-way-on-erc-4626-token-standard-adoption/",
+    "date": "2022-04-08",
+    "thumbnail": "https://cloudfront-us-east-1.images.arcpublishing.com/coindesk/N7ZMZXKDJJDG7EH767HPGOLK7Q.png"
+  },
+  {
+    "name": "Yearn Vaults v3 shipped w/t ERC-4626",
+    "href": "https://medium.com/iearn/yearn-vaults-v3-36ce7c468ca0",
+    "date": "2022-05-02",
+    "thumbnail": "https://miro.medium.com/max/1200/1*CYZe-atKXG5s8vPWmqorMQ.png",
+    "teaser": "Applications built on top of ERC-4626 vaults work with all yield-bearing ERC-4626 vaults, making integration and innovation much easier."
+  },
+  {
+    "name": "How ERC-4626 Could Fuel the Next Wave of DeFi",
+    "href": "https://decrypt.co/99695/how-erc-4626-could-fuel-next-wave-of-defi",
+    "date": "2022-05-11",
+    "thumbnail": "https://cdn.decrypt.co/resize/1024/height/512/wp-content/uploads/2020/12/defi-apps-mass-adoption-gID_6.jpg",
+    "teaser": "Applications built on top of ERC-4626 vaults work with all yield-bearing ERC-4626 vaults, making integration and innovation much easier."
+  },
+  {
+    "name": "A quick primer of ERC-4626A quick primer of ERC-4626",
+    "href": "https://medium.com/mstable/a-quick-primer-of-erc-4626-83a9f62a4272",
+    "date": "2022-05-20",
+    "thumbnail": "https://miro.medium.com/max/4800/1*bUfSL7kxcOXoHecwJo-4rg.jpeg",
+    "teaser": "With the ERC-4626 Tokenized Vault Standard now finalised, we can now focus on developing integrations in a coordinated manner and scale DeFi in a much faster way."
+  },
+  {
+    "name": "ERC4626: A new standard for DeFi applications",
+    "href": "https://chainstack.com/erc4626-a-new-standard-for-defi-applications/",
+    "date": "2022-05-27",
+    "thumbnail": "https://chainstack.com/wp-content/uploads/2022/05/ERC4626-sm.jpeg",
+    "teaser": "ERC4626 is the new standard for tokenized vaults. A building block that extends the ERC20 token and will give a fresh perspective to the DeFi space."
+  },
+  {
+    "name": "DeFi in Ether: Biweekly update on Ethereum DeFi ecosystem vol.54, 24th May — 7th June",
+    "href": "https://medium.com/paradigm-research/defi-in-ether-49-6b-in-defi-mstable-save-gets-erc-4626-upgrade-maker-now-connected-to-starknet-44ca18fb9484",
+    "date": "2022-06-07",
+    "thumbnail": "https://miro.medium.com/max/1200/1*BTVEK72BRkvQTagBgGbSvA.png",
+    "teaser": "DeFi in Ether: $49.6B in DeFi, mStable Save gets ERC-4626 upgrade, Maker now connected to StarkNet, Balancer now live on Optimism, STAKE ←→ GNO token merge is complete, The 0xpo Summit 2022 announced, Bancor 3 roll-out update, and much more! | by Paradigm | Paradigm | Jun, 2022 | Medium"
+  },
+  {
+    "name": "Introducing Apollo: analyzing the usage of ERC4626 across chains",
+    "href": "https://medium.com/chainbound/introducing-apollo-analyzing-the-usage-of-erc4626-across-chains-8668b9b0d83a",
+    "date": "2022-06-15",
+    "thumbnail": "/images/crosschain.jpeg",
+    "teaser": "Since we believe ERC4626 will be a very important building block in future for DeFi protocols, we will focus this analysis around the current state of ERC4626 standard and will be looking at 4 chains: Ethereum, Polygon, Arbitrum and Optimism"
+  },
+  {
+    "name": "Why ERC-4626 could power DeFi’s innovation wave",
+    "href": "https://medium.com/mstable/why-erc-4626-could-power-defis-innovation-wave-588674b0664f",
+    "date": "2022-06-20",
+    "thumbnail": "https://miro.medium.com/max/1200/0*HqJN_h06P9YheVI-.png",
+    "teaser": "Chainbound is a research & development organization focused on public blockchains. This is the first in a series of articles which we believe will shape Chainbound’s mission in the years to come"
+  },
+  {
+    "name": "What is the ERC-4626 token standard?",
+    "href": "https://www.alchemy.com/overviews/erc-4626",
+    "date": "2022-06-21",
+    "thumbnail": "https://assets-global.website-files.com/5f973c97cf5aea614f93a26c/62cc8280a4113062dce12541_Overview-og-1-48.png",
+    "teaser": "Explore how the new ERC-4626 token standard enables Web3 developers creating DeFi applications to build tokenized vaults with increased composability and security."
+  },
+  {
+    "name": "How ERC-4626 can revitalize DeFi",
+    "href": "https://cryplogger.com/erc-4626-how-a-new-standard-can-revitalize-the-defi-segment/",
+    "date": "2022-08-10",
+    "thumbnail": "https://cryplogger.com/wp-content/uploads/2022/08/erc4626_defi-min-1140x570.png",
+    "teaser": "The DeFi ecosystem in particular can be beneficially affected by the new standard of tokens as it's designed to improve communication dApps and expand opportunities for income-generating farming."
+  }
+]
