+++ src/content/videos.json
@@ -1,0 +1,18 @@
+[
+  {
+    "name": "Building with the ERC-4626 Standard",
+    "videoId": "CI6jtEAWHbo"
+  },
+  {
+    "name": "Joey Santoro ERC-4626 deep dive",
+    "videoId": "5KVD7EX6HWQ"
+  },
+  {
+    "name": "Securing ERC4626 Implementations",
+    "videoId": "L8dijE5qhTg"
+  },
+  {
+    "name": "Yearn vault 4626 implementation",
+    "videoId": "urC35PMbpJ4"
+  }
+]
